Add CatalogTransaction abstraction
